create custom module for wide tile tests
