+++ sparse_strips/vello_sparse_tests/snapshots/wide_tile_clip_single_wide_tile.png
