+++ sparse_strips/vello_sparse_tests/snapshots/wide_tile_filter_with_complex_paint_and_wide_tile_shift.png
