+++ sparse_strips/vello_sparse_tests/snapshots/wide_tile_gradient_on_3_wide_tiles.png
