+++ sparse_strips/vello_sparse_tests/snapshots/wide_tile_issue_coarse_batch_in_later_round.png
