+++ sparse_strips/vello_sparse_tests/snapshots/wide_tile_issue_fast_path_strips_and_coarse_batch_in_later_round.png
