+++ sparse_strips/vello_sparse_tests/tests/clip.rs
@@ -18,25 +18,6 @@ use vello_cpu::peniko::{
 };
 use vello_dev_macros::vello_test;
 
-#[vello_test(height = 8)]
-fn nested_clips_within_viewport(ctx: &mut impl Renderer) {
-    const WIDTH: f64 = 100.0;
-    const HEIGHT: f64 = Tile::HEIGHT as f64;
-    const OFFSET: f64 = WIDTH / 3.0;
-
-    let colors = [RED, GREEN, BLUE];
-
-    for (i, color) in colors.iter().enumerate() {
-        let clip_rect = Rect::new((i as f64) * OFFSET, 0.0, WIDTH, HEIGHT);
-        ctx.push_clip_layer(&clip_rect.to_path(0.1));
-        ctx.set_paint(*color);
-        ctx.fill_rect(&Rect::new(0.0, 0.0, WIDTH, HEIGHT));
-    }
-    for _ in colors.iter() {
-        ctx.pop_layer();
-    }
-}
-
 #[vello_test(hybrid_tolerance = 1)]
 fn clip_triangle_with_star(ctx: &mut impl Renderer) {
     let mut triangle_path = BezPath::new();
@@ -365,13 +346,6 @@ fn draw_clipping_outline(ctx: &mut impl Renderer, path: &BezPath) {
 fn clip_exceeding_viewport(ctx: &mut impl Renderer) {
     ctx.push_clip_layer(&Rect::new(0.0, 0.0, 500.0, 10.0).to_path(0.1));
     ctx.fill_rect(&Rect::new(0.0, 0.0, 100.0, 100.0));
-    ctx.pop_layer();
-}
-
-// See <https://github.com/linebender/vello/pull/975#issuecomment-2858372366>
-#[vello_test(no_ref)]
-fn empty_clip_completely_outside_viewport(ctx: &mut impl Renderer) {
-    ctx.push_clip_layer(&Rect::new(300.0, 8.0, 350.0, 48.0).to_path(0.1));
     ctx.pop_layer();
 }
